chore(tests): apply pre-commit formatting and import fixes\n\nApply black/isort fixes to tests/conftest.py (resolve E402/D415).\n\nRefs: #220
